move existing to input-command
