+++ web/src/ee/features/playground/page/components/SaveToPromptButton.tsx
@@ -11,7 +11,7 @@ import {
   InputCommandInput,
   InputCommandItem,
   InputCommandList,
-} from "@/src/components/ui/command";
+} from "@/src/components/ui/input-command";
 import {
   Popover,
   PopoverContent,
