+++ web/src/features/dashboard/components/ModelUsageChart.tsx
@@ -11,7 +11,7 @@ import {
   InputCommandItem,
   InputCommandList,
   InputCommandSeparator,
-} from "@/src/components/ui/command";
+} from "@/src/components/ui/input-command";
 import {
   Popover,
   PopoverContent,
