+++ web/src/features/prompts/components/SetPromptVersionLabels/index.tsx
@@ -8,7 +8,7 @@ import {
   InputCommandGroup,
   InputCommandList,
   InputCommandSeparator,
-} from "@/src/components/ui/command";
+} from "@/src/components/ui/input-command";
 import {
   Popover,
   PopoverContent,
